Set FullyQualifiedNameAttributeName and move functions to external_volumes_gen.go
