+++ pkg/resources/helpers_test.go
@@ -403,596 +403,3 @@ func Test_DataTypeIssue3007DiffSuppressFunc(t *testing.T) {
 		})
 	}
 }
-
-// External volume helper tests
-
-func Test_GetStorageLocationName(t *testing.T) {
-	s3StorageLocationName := "s3Test"
-	s3StorageBaseUrl := "s3://my_example_bucket"
-	s3StorageAwsRoleArn := "arn:aws:iam::123456789012:role/myrole"
-	s3EncryptionKmsKeyId := "123456789"
-
-	gcsStorageLocationName := "gcsTest"
-	gcsStorageBaseUrl := "gcs://my_example_bucket"
-	gcsEncryptionKmsKeyId := "123456789"
-
-	azureStorageLocationName := "azureTest"
-	azureStorageBaseUrl := "azure://123456789.blob.core.windows.net/my_example_container"
-	azureTenantId := "123456789"
-
-	s3StorageLocationA := sdk.S3StorageLocationParams{
-		Name:                 s3StorageLocationName,
-		StorageProvider:      sdk.S3StorageProviderS3,
-		StorageBaseUrl:       s3StorageBaseUrl,
-		StorageAwsRoleArn:    s3StorageAwsRoleArn,
-		StorageAwsExternalId: &s3StorageAwsExternalId,
-		Encryption: &sdk.ExternalVolumeS3Encryption{
-			Type:     sdk.S3EncryptionTypeSseKms,
-			KmsKeyId: &s3EncryptionKmsKeyId,
-		},
-	}
-
-	azureStorageLocationA := sdk.AzureStorageLocationParams{
-		Name:           azureStorageLocationName,
-		StorageBaseUrl: azureStorageBaseUrl,
-		AzureTenantId:  azureTenantId,
-	}
-
-	gcsStorageLocationA := sdk.GCSStorageLocationParams{
-		Name:           gcsStorageLocationName,
-		StorageBaseUrl: gcsStorageBaseUrl,
-		Encryption: &sdk.ExternalVolumeGCSEncryption{
-			Type:     sdk.GCSEncryptionTypeSseKms,
-			KmsKeyId: &gcsEncryptionKmsKeyId,
-		},
-	}
-
-	s3GovStorageLocationA := sdk.S3StorageLocationParams{
-		Name:              s3StorageLocationName,
-		StorageProvider:   sdk.S3StorageProviderS3GOV,
-		StorageBaseUrl:    s3StorageBaseUrl,
-		StorageAwsRoleArn: s3StorageAwsRoleArn,
-	}
-
-	testCases := []struct {
-		Name            string
-		StorageLocation sdk.ExternalVolumeStorageLocation
-		ExpectedName    string
-	}{
-		{
-			Name:            "S3 storage location name succesfully read",
-			StorageLocation: sdk.ExternalVolumeStorageLocation{S3StorageLocationParams: &s3StorageLocationA},
-			ExpectedName:    s3StorageLocationA.Name,
-		},
-		{
-			Name:            "S3GOV storage location name succesfully read",
-			StorageLocation: sdk.ExternalVolumeStorageLocation{S3StorageLocationParams: &s3GovStorageLocationA},
-			ExpectedName:    s3GovStorageLocationA.Name,
-		},
-		{
-			Name:            "GCS storage location name succesfully read",
-			StorageLocation: sdk.ExternalVolumeStorageLocation{GCSStorageLocationParams: &gcsStorageLocationA},
-			ExpectedName:    gcsStorageLocationA.Name,
-		},
-		{
-			Name:            "Azure storage location name succesfully read",
-			StorageLocation: sdk.ExternalVolumeStorageLocation{AzureStorageLocationParams: &azureStorageLocationA},
-			ExpectedName:    azureStorageLocationA.Name,
-		},
-	}
-
-	invalidTestCases := []struct {
-		Name            string
-		StorageLocation sdk.ExternalVolumeStorageLocation
-	}{
-		{
-			Name:            "Empty S3 storage location",
-			StorageLocation: sdk.ExternalVolumeStorageLocation{S3StorageLocationParams: &sdk.S3StorageLocationParams{}},
-		},
-		{
-			Name:            "Empty GCS storage location",
-			StorageLocation: sdk.ExternalVolumeStorageLocation{GCSStorageLocationParams: &sdk.GCSStorageLocationParams{}},
-		},
-		{
-			Name:            "Empty Azure storage location",
-			StorageLocation: sdk.ExternalVolumeStorageLocation{AzureStorageLocationParams: &sdk.AzureStorageLocationParams{}},
-		},
-		{
-			Name:            "Empty storage location",
-			StorageLocation: sdk.ExternalVolumeStorageLocation{},
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.Name, func(t *testing.T) {
-			name, err := resources.GetStorageLocationName(tc.StorageLocation)
-			require.NoError(t, err)
-			assert.Equal(t, tc.ExpectedName, name)
-		})
-	}
-	for _, tc := range invalidTestCases {
-		t.Run(tc.Name, func(t *testing.T) {
-			_, err := resources.GetStorageLocationName(tc.StorageLocation)
-			require.Error(t, err)
-		})
-	}
-}
-
-func Test_GetStorageLocationStorageProvider(t *testing.T) {
-	s3StorageLocationName := "s3Test"
-	s3StorageBaseUrl := "s3://my_example_bucket"
-	s3StorageAwsRoleArn := "arn:aws:iam::123456789012:role/myrole"
-	s3EncryptionKmsKeyId := "123456789"
-
-	gcsStorageLocationName := "gcsTest"
-	gcsStorageBaseUrl := "gcs://my_example_bucket"
-	gcsEncryptionKmsKeyId := "123456789"
-
-	azureStorageLocationName := "azureTest"
-	azureStorageBaseUrl := "azure://123456789.blob.core.windows.net/my_example_container"
-	azureTenantId := "123456789"
-
-	s3StorageLocationA := sdk.S3StorageLocationParams{
-		Name:                 s3StorageLocationName,
-		StorageProvider:      sdk.S3StorageProviderS3,
-		StorageBaseUrl:       s3StorageBaseUrl,
-		StorageAwsRoleArn:    s3StorageAwsRoleArn,
-		StorageAwsExternalId: &s3StorageAwsExternalId,
-		Encryption: &sdk.ExternalVolumeS3Encryption{
-			Type:     sdk.S3EncryptionTypeSseKms,
-			KmsKeyId: &s3EncryptionKmsKeyId,
-		},
-	}
-
-	azureStorageLocationA := sdk.AzureStorageLocationParams{
-		Name:           azureStorageLocationName,
-		StorageBaseUrl: azureStorageBaseUrl,
-		AzureTenantId:  azureTenantId,
-	}
-
-	gcsStorageLocationA := sdk.GCSStorageLocationParams{
-		Name:           gcsStorageLocationName,
-		StorageBaseUrl: gcsStorageBaseUrl,
-		Encryption: &sdk.ExternalVolumeGCSEncryption{
-			Type:     sdk.GCSEncryptionTypeSseKms,
-			KmsKeyId: &gcsEncryptionKmsKeyId,
-		},
-	}
-
-	s3GovStorageLocationA := sdk.S3StorageLocationParams{
-		Name:              s3StorageLocationName,
-		StorageProvider:   sdk.S3StorageProviderS3GOV,
-		StorageBaseUrl:    s3StorageBaseUrl,
-		StorageAwsRoleArn: s3StorageAwsRoleArn,
-	}
-	testCases := []struct {
-		Name                    string
-		StorageLocation         sdk.ExternalVolumeStorageLocation
-		ExpectedStorageProvider sdk.StorageProvider
-	}{
-		{
-			Name:                    "S3 storage provider",
-			StorageLocation:         sdk.ExternalVolumeStorageLocation{S3StorageLocationParams: &s3StorageLocationA},
-			ExpectedStorageProvider: sdk.StorageProviderS3,
-		},
-		{
-			Name:                    "S3GOV storage provider",
-			StorageLocation:         sdk.ExternalVolumeStorageLocation{S3StorageLocationParams: &s3GovStorageLocationA},
-			ExpectedStorageProvider: sdk.StorageProviderS3GOV,
-		},
-		{
-			Name:                    "GCS storage provider",
-			StorageLocation:         sdk.ExternalVolumeStorageLocation{GCSStorageLocationParams: &gcsStorageLocationA},
-			ExpectedStorageProvider: sdk.StorageProviderGCS,
-		},
-		{
-			Name:                    "Azure storage provider",
-			StorageLocation:         sdk.ExternalVolumeStorageLocation{AzureStorageLocationParams: &azureStorageLocationA},
-			ExpectedStorageProvider: sdk.StorageProviderAzure,
-		},
-	}
-
-	invalidTestCases := []struct {
-		Name            string
-		StorageLocation sdk.ExternalVolumeStorageLocation
-	}{
-		{
-			Name:            "Empty S3 storage location",
-			StorageLocation: sdk.ExternalVolumeStorageLocation{S3StorageLocationParams: &sdk.S3StorageLocationParams{}},
-		},
-		{
-			Name:            "Empty GCS storage location",
-			StorageLocation: sdk.ExternalVolumeStorageLocation{GCSStorageLocationParams: &sdk.GCSStorageLocationParams{}},
-		},
-		{
-			Name:            "Empty Azure storage location",
-			StorageLocation: sdk.ExternalVolumeStorageLocation{AzureStorageLocationParams: &sdk.AzureStorageLocationParams{}},
-		},
-		{
-			Name:            "Empty storage location",
-			StorageLocation: sdk.ExternalVolumeStorageLocation{},
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.Name, func(t *testing.T) {
-			storageProvider, err := resources.GetStorageLocationStorageProvider(tc.StorageLocation)
-			require.NoError(t, err)
-			assert.Equal(t, tc.ExpectedStorageProvider, storageProvider)
-		})
-	}
-	for _, tc := range invalidTestCases {
-		t.Run(tc.Name, func(t *testing.T) {
-			_, err := resources.GetStorageLocationName(tc.StorageLocation)
-			require.Error(t, err)
-		})
-	}
-}
-
-var s3StorageAwsExternalId = "1234567890"
-
-func Test_CopySentinelStorageLocation(t *testing.T) {
-	tempStorageLocationName := "terraform_provider_sentinel_storage_location"
-	s3StorageLocationName := "s3Test"
-	s3StorageBaseUrl := "s3://my_example_bucket"
-	s3StorageAwsRoleArn := "arn:aws:iam::123456789012:role/myrole"
-	s3EncryptionKmsKeyId := "123456789"
-
-	gcsStorageLocationName := "gcsTest"
-	gcsStorageBaseUrl := "gcs://my_example_bucket"
-	gcsEncryptionKmsKeyId := "123456789"
-
-	azureStorageLocationName := "azureTest"
-	azureStorageBaseUrl := "azure://123456789.blob.core.windows.net/my_example_container"
-	azureTenantId := "123456789"
-
-	s3StorageLocationA := sdk.S3StorageLocationParams{
-		Name:                 s3StorageLocationName,
-		StorageProvider:      sdk.S3StorageProviderS3,
-		StorageBaseUrl:       s3StorageBaseUrl,
-		StorageAwsRoleArn:    s3StorageAwsRoleArn,
-		StorageAwsExternalId: &s3StorageAwsExternalId,
-		Encryption: &sdk.ExternalVolumeS3Encryption{
-			Type:     sdk.S3EncryptionTypeSseKms,
-			KmsKeyId: &s3EncryptionKmsKeyId,
-		},
-	}
-
-	azureStorageLocationA := sdk.AzureStorageLocationParams{
-		Name:           azureStorageLocationName,
-		StorageBaseUrl: azureStorageBaseUrl,
-		AzureTenantId:  azureTenantId,
-	}
-
-	gcsStorageLocationA := sdk.GCSStorageLocationParams{
-		Name:           gcsStorageLocationName,
-		StorageBaseUrl: gcsStorageBaseUrl,
-		Encryption: &sdk.ExternalVolumeGCSEncryption{
-			Type:     sdk.GCSEncryptionTypeSseKms,
-			KmsKeyId: &gcsEncryptionKmsKeyId,
-		},
-	}
-
-	t.Run("S3 storage location", func(t *testing.T) {
-		storageLocationInput := sdk.ExternalVolumeStorageLocation{S3StorageLocationParams: &s3StorageLocationA}
-		copiedStorageLocation, err := resources.CopySentinelStorageLocation(storageLocationInput)
-		require.NoError(t, err)
-		assert.Equal(t, copiedStorageLocation.S3StorageLocationParams.Name, tempStorageLocationName)
-		assert.Equal(t, copiedStorageLocation.S3StorageLocationParams.StorageProvider, s3StorageLocationA.StorageProvider)
-		assert.Equal(t, copiedStorageLocation.S3StorageLocationParams.StorageBaseUrl, s3StorageLocationA.StorageBaseUrl)
-		assert.Equal(t, copiedStorageLocation.S3StorageLocationParams.StorageAwsRoleArn, s3StorageLocationA.StorageAwsRoleArn)
-		assert.Equal(t, copiedStorageLocation.S3StorageLocationParams.StorageAwsExternalId, s3StorageLocationA.StorageAwsExternalId)
-		assert.Equal(t, copiedStorageLocation.S3StorageLocationParams.Encryption.Type, s3StorageLocationA.Encryption.Type)
-		assert.Equal(t, *copiedStorageLocation.S3StorageLocationParams.Encryption.KmsKeyId, *s3StorageLocationA.Encryption.KmsKeyId)
-	})
-
-	t.Run("GCS storage location", func(t *testing.T) {
-		storageLocationInput := sdk.ExternalVolumeStorageLocation{GCSStorageLocationParams: &gcsStorageLocationA}
-		copiedStorageLocation, err := resources.CopySentinelStorageLocation(storageLocationInput)
-		require.NoError(t, err)
-		assert.Equal(t, copiedStorageLocation.GCSStorageLocationParams.Name, tempStorageLocationName)
-		assert.Equal(t, copiedStorageLocation.GCSStorageLocationParams.StorageBaseUrl, gcsStorageLocationA.StorageBaseUrl)
-		assert.Equal(t, copiedStorageLocation.GCSStorageLocationParams.Encryption.Type, gcsStorageLocationA.Encryption.Type)
-		assert.Equal(t, *copiedStorageLocation.GCSStorageLocationParams.Encryption.KmsKeyId, *gcsStorageLocationA.Encryption.KmsKeyId)
-	})
-
-	t.Run("Azure storage location", func(t *testing.T) {
-		storageLocationInput := sdk.ExternalVolumeStorageLocation{AzureStorageLocationParams: &azureStorageLocationA}
-		copiedStorageLocation, err := resources.CopySentinelStorageLocation(storageLocationInput)
-		require.NoError(t, err)
-		assert.Equal(t, copiedStorageLocation.AzureStorageLocationParams.Name, tempStorageLocationName)
-		assert.Equal(t, copiedStorageLocation.AzureStorageLocationParams.StorageBaseUrl, azureStorageLocationA.StorageBaseUrl)
-		assert.Equal(t, copiedStorageLocation.AzureStorageLocationParams.AzureTenantId, azureStorageLocationA.AzureTenantId)
-	})
-
-	invalidTestCases := []struct {
-		Name            string
-		StorageLocation sdk.ExternalVolumeStorageLocation
-	}{
-		{
-			Name:            "Empty S3 storage location",
-			StorageLocation: sdk.ExternalVolumeStorageLocation{S3StorageLocationParams: &sdk.S3StorageLocationParams{}},
-		},
-		{
-			Name:            "Empty GCS storage location",
-			StorageLocation: sdk.ExternalVolumeStorageLocation{GCSStorageLocationParams: &sdk.GCSStorageLocationParams{}},
-		},
-		{
-			Name:            "Empty Azure storage location",
-			StorageLocation: sdk.ExternalVolumeStorageLocation{AzureStorageLocationParams: &sdk.AzureStorageLocationParams{}},
-		},
-		{
-			Name:            "Empty storage location",
-			StorageLocation: sdk.ExternalVolumeStorageLocation{},
-		},
-	}
-
-	for _, tc := range invalidTestCases {
-		t.Run(tc.Name, func(t *testing.T) {
-			_, err := resources.CopySentinelStorageLocation(tc.StorageLocation)
-			require.Error(t, err)
-		})
-	}
-}
-
-func Test_CommonPrefixLastIndex(t *testing.T) {
-	s3StorageLocationName := "s3Test"
-	s3StorageLocationName2 := "s3Test2"
-	s3StorageBaseUrl := "s3://my_example_bucket"
-	s3StorageAwsRoleArn := "arn:aws:iam::123456789012:role/myrole"
-	s3EncryptionKmsKeyId := "123456789"
-
-	gcsStorageLocationName := "gcsTest"
-	gcsStorageLocationName2 := "gcsTest2"
-	gcsStorageBaseUrl := "gcs://my_example_bucket"
-	gcsEncryptionKmsKeyId := "123456789"
-
-	azureStorageLocationName := "azureTest"
-	azureStorageLocationName2 := "azureTest2"
-	azureStorageBaseUrl := "azure://123456789.blob.core.windows.net/my_example_container"
-	azureTenantId := "123456789"
-
-	s3StorageLocationA := sdk.S3StorageLocationParams{
-		Name:                 s3StorageLocationName,
-		StorageProvider:      sdk.S3StorageProviderS3,
-		StorageBaseUrl:       s3StorageBaseUrl,
-		StorageAwsRoleArn:    s3StorageAwsRoleArn,
-		StorageAwsExternalId: &s3StorageAwsExternalId,
-		Encryption: &sdk.ExternalVolumeS3Encryption{
-			Type:     sdk.S3EncryptionTypeSseKms,
-			KmsKeyId: &s3EncryptionKmsKeyId,
-		},
-	}
-
-	s3StorageLocationB := sdk.S3StorageLocationParams{
-		Name:                 s3StorageLocationName2,
-		StorageProvider:      sdk.S3StorageProviderS3,
-		StorageBaseUrl:       s3StorageBaseUrl,
-		StorageAwsRoleArn:    s3StorageAwsRoleArn,
-		StorageAwsExternalId: &s3StorageAwsExternalId,
-		Encryption: &sdk.ExternalVolumeS3Encryption{
-			Type:     sdk.S3EncryptionTypeSseKms,
-			KmsKeyId: &s3EncryptionKmsKeyId,
-		},
-	}
-
-	azureStorageLocationA := sdk.AzureStorageLocationParams{
-		Name:           azureStorageLocationName,
-		StorageBaseUrl: azureStorageBaseUrl,
-		AzureTenantId:  azureTenantId,
-	}
-
-	azureStorageLocationB := sdk.AzureStorageLocationParams{
-		Name:           azureStorageLocationName2,
-		StorageBaseUrl: azureStorageBaseUrl,
-		AzureTenantId:  azureTenantId,
-	}
-
-	gcsStorageLocationA := sdk.GCSStorageLocationParams{
-		Name:           gcsStorageLocationName,
-		StorageBaseUrl: gcsStorageBaseUrl,
-		Encryption: &sdk.ExternalVolumeGCSEncryption{
-			Type:     sdk.GCSEncryptionTypeSseKms,
-			KmsKeyId: &gcsEncryptionKmsKeyId,
-		},
-	}
-
-	gcsStorageLocationB := sdk.GCSStorageLocationParams{
-		Name:           gcsStorageLocationName2,
-		StorageBaseUrl: gcsStorageBaseUrl,
-		Encryption: &sdk.ExternalVolumeGCSEncryption{
-			Type:     sdk.GCSEncryptionTypeSseKms,
-			KmsKeyId: &gcsEncryptionKmsKeyId,
-		},
-	}
-
-	gcsStorageLocationC := sdk.GCSStorageLocationParams{
-		Name:           "test",
-		StorageBaseUrl: gcsStorageBaseUrl,
-		Encryption: &sdk.ExternalVolumeGCSEncryption{
-			Type:     sdk.GCSEncryptionTypeSseKms,
-			KmsKeyId: &gcsEncryptionKmsKeyId,
-		},
-	}
-
-	s3GovStorageLocationA := sdk.S3StorageLocationParams{
-		Name:              s3StorageLocationName,
-		StorageProvider:   sdk.S3StorageProviderS3GOV,
-		StorageBaseUrl:    s3StorageBaseUrl,
-		StorageAwsRoleArn: s3StorageAwsRoleArn,
-	}
-
-	testCases := []struct {
-		Name           string
-		ListA          []sdk.ExternalVolumeStorageLocation
-		ListB          []sdk.ExternalVolumeStorageLocation
-		ExpectedOutput int
-	}{
-		{
-			Name:           "Two empty lists",
-			ListA:          []sdk.ExternalVolumeStorageLocation{},
-			ListB:          []sdk.ExternalVolumeStorageLocation{},
-			ExpectedOutput: -1,
-		},
-		{
-			Name:           "First list empty",
-			ListA:          []sdk.ExternalVolumeStorageLocation{},
-			ListB:          []sdk.ExternalVolumeStorageLocation{{S3StorageLocationParams: &s3StorageLocationA}},
-			ExpectedOutput: -1,
-		},
-		{
-			Name:           "Second list empty",
-			ListA:          []sdk.ExternalVolumeStorageLocation{{S3StorageLocationParams: &s3StorageLocationA}},
-			ListB:          []sdk.ExternalVolumeStorageLocation{},
-			ExpectedOutput: -1,
-		},
-		{
-			Name:           "Lists with no common prefix - length 1",
-			ListA:          []sdk.ExternalVolumeStorageLocation{{S3StorageLocationParams: &s3StorageLocationA}},
-			ListB:          []sdk.ExternalVolumeStorageLocation{{S3StorageLocationParams: &s3StorageLocationB}},
-			ExpectedOutput: -1,
-		},
-		{
-			Name:           "Lists with no common prefix - length 2",
-			ListA:          []sdk.ExternalVolumeStorageLocation{{S3StorageLocationParams: &s3StorageLocationA}, {AzureStorageLocationParams: &azureStorageLocationA}},
-			ListB:          []sdk.ExternalVolumeStorageLocation{{S3StorageLocationParams: &s3StorageLocationB}, {AzureStorageLocationParams: &azureStorageLocationB}},
-			ExpectedOutput: -1,
-		},
-		{
-			Name:           "Identical lists - length 1",
-			ListA:          []sdk.ExternalVolumeStorageLocation{{S3StorageLocationParams: &s3StorageLocationA}},
-			ListB:          []sdk.ExternalVolumeStorageLocation{{S3StorageLocationParams: &s3StorageLocationA}},
-			ExpectedOutput: 0,
-		},
-		{
-			Name:           "Identical lists - length 2",
-			ListA:          []sdk.ExternalVolumeStorageLocation{{S3StorageLocationParams: &s3StorageLocationA}, {AzureStorageLocationParams: &azureStorageLocationA}},
-			ListB:          []sdk.ExternalVolumeStorageLocation{{S3StorageLocationParams: &s3StorageLocationA}, {AzureStorageLocationParams: &azureStorageLocationA}},
-			ExpectedOutput: 1,
-		},
-		{
-			Name: "Identical lists - length 3",
-			ListA: []sdk.ExternalVolumeStorageLocation{
-				{S3StorageLocationParams: &s3StorageLocationA},
-				{AzureStorageLocationParams: &azureStorageLocationA},
-				{S3StorageLocationParams: &s3GovStorageLocationA},
-			},
-			ListB: []sdk.ExternalVolumeStorageLocation{
-				{S3StorageLocationParams: &s3StorageLocationA},
-				{AzureStorageLocationParams: &azureStorageLocationA},
-				{S3StorageLocationParams: &s3GovStorageLocationA},
-			},
-			ExpectedOutput: 2,
-		},
-		{
-			Name: "Lists with a common prefix - length 3, matching up to and including index 1",
-			ListA: []sdk.ExternalVolumeStorageLocation{
-				{S3StorageLocationParams: &s3StorageLocationA},
-				{AzureStorageLocationParams: &azureStorageLocationA},
-				{GCSStorageLocationParams: &gcsStorageLocationA},
-			},
-			ListB: []sdk.ExternalVolumeStorageLocation{
-				{S3StorageLocationParams: &s3StorageLocationA},
-				{AzureStorageLocationParams: &azureStorageLocationA},
-				{GCSStorageLocationParams: &gcsStorageLocationB},
-			},
-			ExpectedOutput: 1,
-		},
-		{
-			Name: "Lists with a common prefix - length 4, matching up to and including index 2",
-			ListA: []sdk.ExternalVolumeStorageLocation{
-				{S3StorageLocationParams: &s3StorageLocationA},
-				{AzureStorageLocationParams: &azureStorageLocationA},
-				{GCSStorageLocationParams: &gcsStorageLocationA},
-				{GCSStorageLocationParams: &gcsStorageLocationB},
-			},
-			ListB: []sdk.ExternalVolumeStorageLocation{
-				{S3StorageLocationParams: &s3StorageLocationA},
-				{AzureStorageLocationParams: &azureStorageLocationA},
-				{GCSStorageLocationParams: &gcsStorageLocationA},
-				{GCSStorageLocationParams: &gcsStorageLocationC},
-			},
-			ExpectedOutput: 2,
-		},
-		{
-			Name: "Lists with a common prefix - length 4, matching up to and including index 1",
-			ListA: []sdk.ExternalVolumeStorageLocation{
-				{S3StorageLocationParams: &s3StorageLocationA},
-				{AzureStorageLocationParams: &azureStorageLocationA},
-				{GCSStorageLocationParams: &gcsStorageLocationA},
-				{GCSStorageLocationParams: &gcsStorageLocationC},
-			},
-			ListB: []sdk.ExternalVolumeStorageLocation{
-				{S3StorageLocationParams: &s3StorageLocationA},
-				{AzureStorageLocationParams: &azureStorageLocationA},
-				{GCSStorageLocationParams: &gcsStorageLocationB},
-				{GCSStorageLocationParams: &gcsStorageLocationC},
-			},
-			ExpectedOutput: 1,
-		},
-		{
-			Name: "Lists with a common prefix - different lengths, matching up to and including index 1 (last index of shorter list)",
-			ListA: []sdk.ExternalVolumeStorageLocation{
-				{S3StorageLocationParams: &s3StorageLocationA},
-				{AzureStorageLocationParams: &azureStorageLocationA},
-				{GCSStorageLocationParams: &gcsStorageLocationA},
-			},
-			ListB: []sdk.ExternalVolumeStorageLocation{
-				{S3StorageLocationParams: &s3StorageLocationA},
-				{AzureStorageLocationParams: &azureStorageLocationA},
-			},
-			ExpectedOutput: 1,
-		},
-		{
-			Name: "Lists with a common prefix - different lengths, matching up to and including index 2",
-			ListA: []sdk.ExternalVolumeStorageLocation{
-				{S3StorageLocationParams: &s3StorageLocationA},
-				{AzureStorageLocationParams: &azureStorageLocationA},
-				{S3StorageLocationParams: &s3StorageLocationB},
-				{GCSStorageLocationParams: &gcsStorageLocationA},
-				{GCSStorageLocationParams: &gcsStorageLocationB},
-				{AzureStorageLocationParams: &azureStorageLocationB},
-			},
-			ListB: []sdk.ExternalVolumeStorageLocation{
-				{S3StorageLocationParams: &s3StorageLocationA},
-				{AzureStorageLocationParams: &azureStorageLocationA},
-				{S3StorageLocationParams: &s3StorageLocationB},
-				{GCSStorageLocationParams: &gcsStorageLocationB},
-				{AzureStorageLocationParams: &azureStorageLocationB},
-			},
-			ExpectedOutput: 2,
-		},
-		{
-			Name:           "Empty S3 storage location",
-			ListA:          []sdk.ExternalVolumeStorageLocation{{S3StorageLocationParams: &s3StorageLocationA}},
-			ListB:          []sdk.ExternalVolumeStorageLocation{{S3StorageLocationParams: &sdk.S3StorageLocationParams{}}},
-			ExpectedOutput: -1,
-		},
-		{
-			Name:           "Empty GCS storage location",
-			ListA:          []sdk.ExternalVolumeStorageLocation{{S3StorageLocationParams: &s3StorageLocationA}},
-			ListB:          []sdk.ExternalVolumeStorageLocation{{GCSStorageLocationParams: &sdk.GCSStorageLocationParams{}}},
-			ExpectedOutput: -1,
-		},
-		{
-			Name:           "Empty Azure storage location",
-			ListA:          []sdk.ExternalVolumeStorageLocation{{S3StorageLocationParams: &s3StorageLocationA}},
-			ListB:          []sdk.ExternalVolumeStorageLocation{{AzureStorageLocationParams: &sdk.AzureStorageLocationParams{}}},
-			ExpectedOutput: -1,
-		},
-		{
-			Name:           "Empty storage location",
-			ListA:          []sdk.ExternalVolumeStorageLocation{{S3StorageLocationParams: &s3StorageLocationA}},
-			ListB:          []sdk.ExternalVolumeStorageLocation{{}},
-			ExpectedOutput: -1,
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.Name, func(t *testing.T) {
-			commonPrefixLastIndex, err := resources.CommonPrefixLastIndex(tc.ListA, tc.ListB)
-			require.NoError(t, err)
-			assert.Equal(t, tc.ExpectedOutput, commonPrefixLastIndex)
-		})
-	}
-}
